Youyou zhang

+++ backend/routes/taskRoute.js
@@ -1,6 +1,12 @@
+const task = require('../models/task');
+
 var express = require('express'),
     router = express.Router(),
-    mongoose = require('mongoose');
+    mongoose = require('mongoose'),
+    taskController = require('../controller/taskController'),
+    tagController = require('../controller/tagController'),
+    collectionController = require('../controller/collectionController'),
+    toId = mongoose.Types.ObjectId;
 
 require('../models/collection');
 require('../models/tag');
@@ -26,14 +32,79 @@ module.exports = (router) => {
             assignedCollection: req.body.assignedCollection
         });
         console.log("! newTask: "+task);
-        task.save().then(doc =>{
+        task.save().then(async doc =>{
             res.status(201).json({
                 message: "Status: Create Success",
                 data: doc
             });
+
+            //after task created, add taskID to given collectionID
+            await collectionController.addTask(req.body.assignedCollection, doc._id);
+
+            //after task created, add taskID to given tagID
+            await tagController.addTask(req.body.tag, doc._id);
         }).catch((err)=>{
             next(err);
-        })        
+        })    
+    });
+
+    //////////////////////////PUT/////////////////////////
+    taskRouteID.put(async(req, res, next)=>{
+        console.log('===in putTask===');
+        console.log("! req.body: "+JSON.stringify(req.body));
+        console.log("! req.params: "+JSON.stringify(req.params.id));
+
+
+        const id=req.params.id;
+        const name = req.body.name? (req.body.name): null;
+        var nameParam; if(name) nameParam=(name); console.log(nameParam);
+        
+        const tag = req.body.tag? (req.body.tag): null;
+        var tagParam; var oldTagId;
+        if(tag) {
+            tagParam=(tag); 
+            console.log(tagParam);
+            
+            oldTagId = await taskController.getTagIDFromTaskID(id);
+        }
+        const description = req.body.description? (req.body.description): null;
+        var descriptionParam; if(description) descriptionParam=toId(description); console.log(descriptionParam);
+
+        const assignedCollection = req.body.assignedCollection? (req.body.assignedCollection): null;
+        var assignedCollectionParam; if(assignedCollection) assignedCollectionParam=toId(assignedCollection); console.log(assignedCollectionParam);
+
+ 
+        const doc = await taskModel.findOneAndUpdate(
+            { _id: id },
+            { 
+                name: nameParam,
+                tag: tagParam,
+                description: descriptionParam,
+                assignedCollection: assignedCollectionParam
+            }, 
+            { new: true }
+        ).catch(next);
+        console.log("! after update: "+(doc)); 
+        
+        if(doc){
+            res.status(201).json({
+                message: "Status: Update Success",
+                data: doc
+            });
+
+            if(tagParam){
+                //after task created, add taskID to given new tagID
+                await tagController.addTask(req.body.tag, doc._id);
+
+                //goto old tagID, delete taskID from alltask[]
+                await tagController.deleteTask(oldTagId, doc._id);
+            }
+        }  
+        else{
+            let err = new Error('Status: Update Failed');
+            err.code = 404;
+            next(err); 
+        };    
     });
 
     //////////////////////////////GET//////////////////////////////////
@@ -43,8 +114,7 @@ module.exports = (router) => {
         const alltask=await taskModel.find();
         console.log("! get all task: "+alltask);
 
-        if(alltask)
-        {
+        if(alltask) {
             res.status(200).json({
                 message: "Status: Get Success",
                 data: alltask
@@ -62,7 +132,18 @@ module.exports = (router) => {
         console.log("! req.body: "+JSON.stringify(req.body));
         console.log("! req.params: "+JSON.stringify(req.params));
         
+        //1 find todelete task CollID + save taskID
         let toDeleteTaskID = req.params.id;
+        if( !(await taskController.isTaskIDExist(toDeleteTaskID)) ){
+            let err = new Error('Status: Delete Task Failed. Given task ID not found');
+            err.code = 404;
+            next(err); 
+            return;
+        }
+        let collectionId = await taskController.getCollIDFromTaskID(toDeleteTaskID);
+        let tagId = await taskController.getTagIDFromTaskID(toDeleteTaskID);
+
+        //2 delete a task
         taskModel.findByIdAndDelete(toDeleteTaskID)
             .then(result =>{
                 console.log("! result: "+result);
@@ -71,7 +152,13 @@ module.exports = (router) => {
                     message: "Status: Delete Success",
                     data: result
                 });
-            }).catch(next);     
+            }).catch(next);   
+            
+        //3 goto CollID, delete taskID from alltask[]
+        await collectionController.deleteTask(collectionId, toDeleteTaskID);
+
+        //4 goto tagID, delete taskID from alltask[]
+        await tagController.deleteTask(tagId, toDeleteTaskID);
     });
